Crate pokemon_rust

Modules

audio
Types related to BGM/SFX playback.

common
Contains common types and functions used throughout the entire game.

config
Types related to the main configuration file (settings.ron).

constants
Common constants used throughout the entire game.

entities
Contains the "actors" of the game.

events
Contains every possible game event. Events can trigger animations, textboxes, warps and much more.

lua
Integration with Lua through FFI. Contains Rust functions that can be called from Lua code and the required plumbing to make this communication work.

map
Controls pretty much everything related to maps in the overworld: which maps are loaded, where are all the characters and items, when are connected maps loaded and much more.

states
The possible states of the game.

systems
The systems present in the game that operate on its entities and resources. Although a big part of the logic is actually encoded in events, there are some important systems that help them do their job.

Functions

main