Module pokemon_rust::common

Contains common types and functions used throughout the entire game.

Structs

AssetTracker
A wrapper around a progress counter that is inserted into the world as a resource. This allows different parts of the game to use the same progress counter.

CommonResources
A cache of diverse resources that are used extensively by different parts of the game.

Enums

Direction
A two-dimensional, four-axis direction enum.

Constants

_IMPL_DESERIALIZE_FOR_Direction

_IMPL_SERIALIZE_FOR_Direction

Functions

get_character_sprite_index_from_direction
Returns the appropriate sprite index to use for a standing character that is facing a given direction.

get_direction_offset
Returns a pair of coordinates (x, y) representing a given direction, expressed in a given type.

load_full_texture_sprite_sheet
Loads a texture from a given image filename, and creates a spritesheet to it representing the entire texture.

load_sprite_sheet
Loads a texture + spritesheet from given image/ron filenames.