Module pokemon_rust::entities::map::conversions

Functions

get_reference_point_from_tile
Given the position of a tile in both Map Coordinates and World Coordinates, calculates the reference point of its map.

map_to_world_coordinates
Given the position of a tile in Map Coordinates and the reference point of its map, calculates the position of the tile in World Coordinates.

player_to_map_coordinates
Given the position of a player and the reference point of its map, calculates the position of its tile in Map Coordinates.

world_to_map_coordinates
Given the position of a tile in World Coordinates and the reference point of its map, calculates the position of the tile in Map Coordinates.