Trait pokemon_rust::events::GameEvent

pub trait GameEvent {
    fn boxed_clone(&self) -> Box<dyn GameEvent + Sync + Send>;

    fn get_execution_conditions(&self) -> ExecutionConditions;

    fn start(&mut self, world: &mut World);

    fn tick(&mut self, world: &mut World, disabled_inputs: bool);

    fn is_complete(&self, world: &mut World) -> bool;
}

A trait representing a game event.

Required methods

fn boxed_clone(&self) -> Box<dyn GameEvent + Sync + Send>

Returns an arbitrary GameEvent inside a Box. This is basically a trait object-safe version of Clone used specifically for game events.

fn get_execution_conditions(&self) -> ExecutionConditions

Returns the execution conditions of this event.

fn start(&mut self, world: &mut World)

Starts this event, assuming that its execution conditions hold.

fn tick(&mut self, world: &mut World, disabled_inputs: bool)

Ticks this event, making it potentially get closer to completion.

fn is_complete(&self, world: &mut World) -> bool

Checks if this event has been completed.

Loading content...