Module pokemon_rust::lua

Integration with Lua through FFI. Contains Rust functions that can be called from Lua code and the required plumbing to make this communication work.

Modules

events

npc

polymorphic_container

Structs

ExecutionContext

Enums

LuaScriptError

Constants

LUA

Functions

get_lua_type_name
Returns a string represention of a Lua type. This is a copy of Value::type_name(), which for some reason is private...

run_lua_script

run_script

run_with_native_functions