Module pokemon_rust::map

Controls pretty much everything related to maps in the overworld: which maps are loaded, where are all the characters and items, when are connected maps loaded and much more.

Re-exports

pub use self::coordinates::CoordinateSystem;

pub use self::coordinates::MapCoordinates;

pub use self::coordinates::PlayerCoordinates;

pub use self::coordinates::WorldCoordinates;

pub use self::conversions::map_to_world_coordinates;

pub use self::load_map::change_player_tile;

pub use self::load_map::initialise_map;

pub use self::load_map::interact_with_npc;

pub use self::load_map::prepare_warp;

pub use self::map::GameAction;

pub use self::map::GameActionKind;

pub use self::map::GameScript;

pub use self::map::GameScriptParameters;

pub use self::map::MapConnection;

pub use self::map::MapScript;

pub use self::map::MapScriptKind;

pub use self::map::Tile;

pub use self::tile_data_builder::PreparedTileDataBuilder;

pub use self::tile_data_builder::TileDataBuilder;

Modules

conversions

coordinates

load_map

map

serializable_map

tile_data_builder

Structs

CharacterData

MapHandler

MapId

TileData
A global way to refer to a tile.

ValidatedGameAction