Module pokemon_rust::systems

The systems present in the game that operate on its entities and resources. Although a big part of the logic is actually encoded in events, there are some important systems that help them do their job.

Re-exports

pub use animation_system::AnimationSystem;

pub use audio_system::AudioSystem;

pub use character_movement_system::CharacterMovementSystem;

pub use map_change_announcement_system::MapChangeAnnouncementSystem;

pub use npc_interaction_system::NpcInteractionSystem;

pub use player_input_system::PlayerInputSystem;

pub use text_system::TextSystem;

Modules

animation_system
A system for animating anything that has an AnimationTable.

audio_system
A system for background music playback. Reads the next BGM from Music and plays it whenever it changes or ends. Affected by GameConfig::play_bgm.

character_movement_system
A system responsible for executing character movements.

map_change_announcement_system
A system for animating map change announcements.

npc_interaction_system
A system for processing interactions with NPCs once they're no longer moving.

player_input_system
A system responsible for reading the player's inputs and processing them appropriately.

text_system
A system responsible for animating a text box on the screen.